refactor Info
