+++ src/Juvix/Compiler/Core/Evaluator.hs
@@ -13,9 +13,9 @@ import GHC.Show as S
 import Juvix.Compiler.Core.Data.InfoTable
 import Juvix.Compiler.Core.Error
 import Juvix.Compiler.Core.Extra
+import Juvix.Compiler.Core.Info qualified as Info
+import Juvix.Compiler.Core.Info.NoDisplayInfo
 import Juvix.Compiler.Core.Language
-import Juvix.Compiler.Core.Language.Info qualified as Info
-import Juvix.Compiler.Core.Language.Info.NoDisplayInfo
 import Juvix.Compiler.Core.Pretty
 
 data EvalError = EvalError
