+++ src/Juvix/Compiler/Core/Extra/Base.hs
@@ -2,9 +2,9 @@ module Juvix.Compiler.Core.Extra.Base where
 
 import Data.Functor.Identity
 import Data.List qualified as List
+import Juvix.Compiler.Core.Info qualified as Info
+import Juvix.Compiler.Core.Info.BinderInfo
 import Juvix.Compiler.Core.Language
-import Juvix.Compiler.Core.Language.Info qualified as Info
-import Juvix.Compiler.Core.Language.Info.BinderInfo
 
 {------------------------------------------------------------------------}
 {- functions on Type -}
