+++ src/Juvix/Compiler/Core/Extra/Info.hs
@@ -2,10 +2,10 @@ module Juvix.Compiler.Core.Extra.Info where
 
 import Juvix.Compiler.Core.Extra.Base
 import Juvix.Compiler.Core.Extra.Recursors
+import Juvix.Compiler.Core.Info qualified as Info
+import Juvix.Compiler.Core.Info.LocationInfo
+import Juvix.Compiler.Core.Info.NameInfo
 import Juvix.Compiler.Core.Language
-import Juvix.Compiler.Core.Language.Info qualified as Info
-import Juvix.Compiler.Core.Language.Info.LocationInfo
-import Juvix.Compiler.Core.Language.Info.NameInfo
 
 mapInfo :: (Info -> Info) -> Node -> Node
 mapInfo f = umap (modifyInfo f)
