+++ src/Juvix/Compiler/Core/Translation/FromSource.hs
@@ -10,12 +10,12 @@ import Data.List qualified as List
 import Juvix.Compiler.Core.Data.InfoTable
 import Juvix.Compiler.Core.Data.InfoTableBuilder
 import Juvix.Compiler.Core.Extra.Base
+import Juvix.Compiler.Core.Info qualified as Info
+import Juvix.Compiler.Core.Info.BinderInfo as BinderInfo
+import Juvix.Compiler.Core.Info.BranchInfo as BranchInfo
+import Juvix.Compiler.Core.Info.LocationInfo as LocationInfo
+import Juvix.Compiler.Core.Info.NameInfo as NameInfo
 import Juvix.Compiler.Core.Language
-import Juvix.Compiler.Core.Language.Info qualified as Info
-import Juvix.Compiler.Core.Language.Info.BinderInfo as BinderInfo
-import Juvix.Compiler.Core.Language.Info.BranchInfo as BranchInfo
-import Juvix.Compiler.Core.Language.Info.LocationInfo as LocationInfo
-import Juvix.Compiler.Core.Language.Info.NameInfo as NameInfo
 import Juvix.Compiler.Core.Transformation.Eta
 import Juvix.Compiler.Core.Translation.FromSource.Lexer
 import Juvix.Parser.Error
